DRC-1997: Fixed long/double conversion error. Does not fixbamboo.

+++ RobotDataCommunication/src/us/ihmc/multicastLogDataProtocol/control/VariableChangeRequest.java
@@ -3,5 +3,5 @@ package us.ihmc.multicastLogDataProtocol.control;
 public class VariableChangeRequest
 {
    public int variableID;
-   public long requestedValue;
+   public double requestedValue;
 }
